add lopp func in helpers.py update globalvariables.py comment logger
